Remove check about XML timestamp
For a reason that I don't explain, the XML update timestamp is roughly a
day behind the source collection timestamps.

https://bugzilla.mozilla.org/show_bug.cgi?id=1675807

This XML is used by Firefox < 68, and although it's still live, we don't support it anymore and don't want to spend human resources for this.

Let's remove this failing check that prevents us from concentrating on
real stuff.